Clase Main propiamente definida, que no hecha
