+++ RecetApp/src/dad/recetapp/tests/TestInstruccionDialog.java
@@ -1,4 +1,4 @@
-package dad.recetapp;
+package dad.recetapp.tests;
 
 import dad.recetapp.ui.ItemDialogFactory;
 import dad.recetapp.ui.controllers.InstruccionDialogController;
